// front-end/src/common/interfaces.ts
export interface Point {
    x: number;
    y: number;
}

export interface ImageState {
    imageUrl: string;
    isChecked: boolean;
    isBase?: boolean;
}
export interface IImage {
    imageName: string;
    imageUrl: string;
}

export interface Color {
    red: number;
    green: number;
    blue: number;
}